Add a module structure and setup files.
