+++ .gitignore
@@ -426,3 +426,7 @@ dmypy.json
 # Cython debug symbols
 cython_debug/
 
+### Custom template
+
+# Ignore CloudFormation Template
+template.yaml
