+++ setup.cfg
@@ -1,0 +1,10 @@
+[metadata]
+name = vpc-cli
+version = 0.0.1
+
+[options]
+packages = vpc_cli
+
+[options.entry_points]
+console_scripts =
+    vpc-cli = vpc_cli.main:main
