+++ vpc_cli/main.py
@@ -1,0 +1,9 @@
+from vpc_cli.command import Command
+
+
+def main():
+    Command()
+
+
+if __name__ == '__main__':
+    main()
