+++ vpc_cli/print_table.py
@@ -1,0 +1,110 @@
+from prettytable import PrettyTable
+
+
+class PrintTable:
+    table = None
+
+    def __init__(self):
+        self.table = PrettyTable()
+        self.table.set_style(15)
+
+    def print_vpc(self, region, vpc):
+        self.table.clear()
+        self.table.title = 'VPC'
+        self.table.field_names = ['Region', 'Name', 'CIDR']
+        self.table.add_row([region, vpc['name'], vpc['cidr']])
+        print(self.table)
+
+    def print_subnets(
+            self,
+            public_subnet=None,
+            private_subnet=None,
+            protected_subnet=None,
+            public_rtb=None,
+            private_rtb=None,
+            protected_rtb=None
+    ):
+        self.table.clear()
+        self.table.title = 'Subnets'
+        self.table.field_names = ['AZ', 'Name', 'CIDR', 'Route Table']
+
+        if public_subnet:
+            for subnet in public_subnet:
+                self.table.add_row([subnet['az'], subnet['name'], subnet['cidr'], public_rtb])
+
+        if private_subnet:
+            for subnet in private_subnet:
+                route_table_name = next((item for item in private_rtb if item['subnet'] == subnet['name']))['name']
+                self.table.add_row([subnet['az'], subnet['name'], subnet['cidr'], route_table_name])
+
+        if protected_subnet:
+            for subnet in protected_subnet:
+                self.table.add_row([subnet['az'], subnet['name'], subnet['cidr'], protected_rtb])
+
+        print(self.table)
+
+    def print_route_tables(
+            self,
+            public_rtb=None,
+            private_rtb=None,
+            protected_rtb=None,
+            igw=None
+    ):
+        self.table.clear()
+        self.table.title = 'Route Tables'
+        self.table.field_names = ['Type', 'Name', 'Gateway']
+
+        if public_rtb:
+            self.table.add_row(['Public', public_rtb, igw])
+
+        if private_rtb:
+            for rtb in private_rtb:
+                self.table.add_row(['Private', rtb['name'], rtb['nat']])
+
+        if protected_rtb:
+            self.table.add_row(['Protected', protected_rtb, 'None'])
+
+        print(self.table)
+
+    def print_igw(
+            self,
+            igw='None'
+    ):
+        self.table.clear()
+        self.table.title = 'Internet Gateway'
+        self.table.field_names = ['Name']
+        self.table.add_row([igw])
+
+        print(self.table)
+
+    def print_nat(
+            self,
+            nat=None,
+    ):
+        self.table.clear()
+        self.table.title = 'NAT Gateways'
+        self.table.field_names = ['Name', 'Elastic IP', 'Subnet']
+
+        if nat:
+            for nat_gw in nat:
+                self.table.add_row([nat_gw['name'], nat_gw['eip'], nat_gw['subnet']])
+        else:
+            self.table.add_row(['None', 'None', 'None'])
+
+        print(self.table)
+
+    def print_s3_ep(
+            self,
+            s3_gateway_ep=None,
+    ):
+        self.table.clear()
+        self.table.title = 'S3 Gateway Endpoint'
+        self.table.field_names = ['Name', 'Route Table']
+
+        if s3_gateway_ep['required']:
+            for route_table in s3_gateway_ep['route-table']:
+                self.table.add_row([s3_gateway_ep['name'], route_table])
+        else:
+            self.table.add_row(['None', 'None'])
+
+        print(self.table)
